+++ plugin/web/extensions/sidebar/utils/dataset.ts
@@ -5,7 +5,10 @@ import {
   FieldComponent,
   SwitchGroup,
 } from "../core/components/content/common/DatasetCard/Field/Fields/types";
-import { DataCatalogItem, Template } from "../core/types";
+import { Data, DataCatalogItem, Template } from "../core/types";
+import { RawDataCatalogItem } from "../modals/datacatalog/api/api";
+
+import { generateID } from ".";
 
 export const getActiveFieldIDs = (
   components?: FieldComponent[],
@@ -58,4 +61,133 @@ export const getDefaultDataset = (dataset?: DataCatalogItem | UserDataItem) => {
   if (dataset.config?.data && dataset.config?.data?.length > 0) {
     return dataset.config.data[0];
   }
+};
+
+export const splitComponents = (
+  activeIDs: string[],
+  dataset: DataCatalogItem,
+  templates?: Template[],
+) => {
+  const flattenedComponents = flattenComponents(dataset.components, templates);
+  const inactiveComponents = flattenedComponents
+    ?.filter(c => !activeIDs.find(id => id === c.id))
+    .map(c => {
+      if (c.type === "switchDataset" && !c.cleanseOverride) {
+        c.cleanseOverride = {
+          data: {
+            url: dataset.config?.data?.[0].url,
+            time: {
+              updateClockOnLoad: false,
+            },
+          },
+        };
+      }
+      return c;
+    });
+  const activeComponents = flattenedComponents
+    ?.filter(c => !!activeIDs.find(id => id === c.id))
+    .map(c => {
+      if (c.type === "switchDataset" && !c.cleanseOverride) {
+        c.cleanseOverride = {
+          data: {
+            url: dataset.config?.data?.[0].url,
+            time: {
+              updateClockOnLoad: false,
+            },
+          },
+        };
+      }
+      return c;
+    });
+  return {
+    activeComponents,
+    inactiveComponents,
+  };
+};
+
+export const processComponentsToSave = (components?: FieldComponent[], templates?: Template[]) => {
+  if (!components) return;
+  return components?.map((c: any) => {
+    const newComp = Object.assign({}, c);
+    if (newComp.type === "template" && newComp.components) {
+      newComp.components =
+        templates
+          ?.find(t => t.id === newComp.templateID)
+          ?.components?.map(c => {
+            return { ...c, userSettings: undefined };
+          }) ?? [];
+    }
+    return { ...newComp, userSettings: undefined };
+  });
+};
+
+export const convertDatasetToData = (dataset: DataCatalogItem, templates?: Template[]): Data => {
+  return {
+    dataID: dataset.dataID,
+    public: dataset.public,
+    components: processComponentsToSave(dataset.components, templates),
+  };
+};
+
+export const newItem = (ri: RawDataCatalogItem): DataCatalogItem => {
+  return {
+    ...ri,
+    dataID: ri.id,
+    public: false,
+    visible: true,
+  };
+};
+
+export const handleDataCatalogProcessing = (
+  catalog: (DataCatalogItem | RawDataCatalogItem)[],
+  savedData?: Data[],
+): DataCatalogItem[] =>
+  catalog.map(item => {
+    if (!savedData) return newItem(item);
+
+    const savedData2 = savedData.find(d => d.dataID === ("dataID" in item ? item.dataID : item.id));
+    if (savedData2) {
+      return {
+        ...item,
+        ...savedData2,
+        visible: true,
+        selectedGroup: getDefaultGroup(savedData2.components),
+      };
+    } else {
+      return newItem(item);
+    }
+  });
+
+export const processDatasetToAdd = (
+  dataset: DataCatalogItem | UserDataItem,
+  templates?: Template[],
+): DataCatalogItem => {
+  const datasetToAdd = { ...dataset };
+
+  datasetToAdd.selectedGroup = getDefaultGroup(datasetToAdd.components, templates);
+  datasetToAdd.selectedDataset = getDefaultDataset(datasetToAdd);
+
+  if (!dataset.components?.length) {
+    const defaultTemplate = templates?.find(t =>
+      dataset.type2
+        ? t.name.includes(dataset.type2)
+        : dataset.type
+        ? t.name.includes(dataset.type)
+        : undefined,
+    );
+    if (defaultTemplate && !datasetToAdd.components) {
+      datasetToAdd.components = [
+        {
+          id: generateID(),
+          type: "template",
+          templateID: defaultTemplate.id,
+          userSettings: {
+            components: defaultTemplate.components,
+          },
+        },
+      ];
+      datasetToAdd.selectedGroup = getDefaultGroup(defaultTemplate.components);
+    }
+  }
+  return datasetToAdd as DataCatalogItem;
 };
